[refactor/TK-253] BaseIntegrationTest 위치 변경
